move test node to tests
