+++ src/lib.rs
@@ -46,10 +46,6 @@ pub mod rpc;
 pub mod store;
 pub mod util;
 
-#[cfg(feature = "rpc")]
-#[cfg_attr(iroh_docsrs, doc(cfg(feature = "rpc")))]
-pub mod node;
-
 use bao_tree::BlockSize;
 pub use iroh_base::hash::{BlobFormat, Hash, HashAndFormat};
